+++ includes/class-customize.php
@@ -6,7 +6,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    1.0.0
- * @version  1.9.1
+ * @version  2.1.0
  *
  * Contents:
  *
@@ -101,7 +101,7 @@ class Invoices_Customize {
 		 * Theme customizer options
 		 *
 		 * @since    1.0.0
-		 * @version  1.9.1
+		 * @version  2.1.0
 		 *
 		 * @param  object $wp_customize  WP customizer object.
 		 */
@@ -233,7 +233,8 @@ class Invoices_Customize {
 					$wp_customize->add_panel(
 						'theme_options',
 						array(
-							'title' => esc_html__( 'Theme Options', '_invoices' ),
+							'title'    => esc_html__( 'Theme Options', '_invoices' ),
+							'priority' => 0,
 						)
 					);
 
@@ -309,28 +310,6 @@ class Invoices_Customize {
 							array(
 								'label'       => esc_html__( 'Default exchange TO currency', '_invoices' ),
 								'description' => esc_html__( 'Set this to your accounting currency - will be used for exchange calculation in dual currency invoices.', '_invoices' ),
-								'section'     => 'currency',
-								'type'        => 'text',
-								'priority'    => ++$priority,
-							)
-						);
-
-					// Option: currency/api_fixer_access_key
-
-						$wp_customize->add_setting(
-							'api_fixer_access_key',
-							array(
-								'default'           => '',
-								'transport'         => $transport,
-								'sanitize_callback' => 'sanitize_text_field',
-							)
-						);
-
-						$wp_customize->add_control(
-							'api_fixer_access_key',
-							array(
-								'label'       => esc_html__( 'Fixer.io API access key', '_invoices' ),
-								'description' => esc_html__( 'Get free access key from https://fixer.io/', '_invoices' ),
 								'section'     => 'currency',
 								'type'        => 'text',
 								'priority'    => ++$priority,
@@ -438,6 +417,7 @@ class Invoices_Customize {
 							'count_invoices_per_month',
 							array(
 								'label'       => esc_html__( 'Max number of invoices created monthly', '_invoices' ),
+								'description' => esc_html__( 'This is used to calculate `posts_per_page` query value.', '_invoices' ),
 								'section'     => 'invoice',
 								'type'        => 'number',
 								'priority'    => ++$priority,